Create BundlerAudit module and nest everything else underneath
