+++ bin/bundler-audit
@@ -10,6 +10,6 @@ else
   engine_config = {}
 end
 
-CC::Engine::BundlerAudit.new(
+CC::Engine::BundlerAudit::Analyzer.new(
   directory: "/code", engine_config: engine_config, io: STDOUT
 ).run
